#3 Turning "helpings" into "ugly."
We wanted for a long time to create a project, called "ugly," to group utilities
fo helping automated testing. (The name is "ugly" because some utilities are
going to do not very elegant things, such as waiting for ports to be listening
or down. Indeed, the main motivation are the functions from
"confeitaria.server.waiters."

Since "ugly" main appeal would be testing, and "helpings" is only about testing,
then it would be better to merge both projects. Since "ugly" does not exist,
"merging" is just renaming ;)