#15 Make server span a new thread when used as context manager.
This way, one can give the server to a with statement and then proceed with any
tests needed. Once all tests are done (and the with block ends) the server stops
itself.

(At first I tried to use multiprocessing but the SocketServer.TCPServer
constructor already allocates the port; since the parent process allocated it,
the child process could not start the server. So I dropped it in favor to use
threading.)