#17 Check whether the server is up before shutting down.
If the server is not started up, do not shut it down.

Now, however, we would be prone to some race condition: the server is going to
shut down and, just after deciding it is not initialized, it starts up. So, it
will be initialized and then the shut down will not happen, letting the port
reserved. That's why we created a lock to the server: it will not shut down
until the starting up is done and vice versa.

As a bonus, we are not waiting for an arbitrary time in __enter__() anymore: whe
can just acquire the lock. I am *almost* liking working with threads.