#14 Move wait_server_up() to ugly.
This function was copied from another project, Confeitaria, more specifically
from the confeitaria.server.waiters module (cf. http://tinyurl.com/nsv5cdy).

To have tests proved a very smart move: here I just discovered that there is not
timeout for refused connections: the socket raises the exception immediately.
That's why there is a time.time() in the except block that was not present in
the Confeitaria version.