#22 Save the raised exception.
How we will use it in practice is yet to be defined. In general, a failure in
the spawned process results from a failure in the original process. This means
that an exception was already raised by the original process and, if it is not
treated, the original process finishes before any treatment to be done at
pc.exceptions.

Nonetheless, this is a very important feature to have. It is dastardly annoying
to not have ANY way to check for exceptions from the spawned process.