#23 Test for "conversation" with process.
Communicating with the process from the context is a common need. Here we show
how one is expected to do it in a more succinct (although not necessarily
simpler) way.

If the function given to the process context is a generator function, then the
generator can expect values from the original process. Also, the original
process will have the values yielded by the generator.