#23 Refactor IPC out of Process Context.
There was this feeling that process context was doing too much. We could isolate
all queue juggling into a different class, Conversation. It seems to be more
readable now, although the class still seems strange.

Nonetheless, the change seems reasonable, not least because we discovered we had
a bad API contract. ProcessContext expected to get a list with the exceptions
not threated by the process - but only one could be there, there is no way of
not treating two or more excpetions. So now we dropped the pc.exceptions list in
favor of pc.exception (which can be None or an exception). Neat, uh?