#27 Turn waiter's timeout arg into total expected timeout.
If a waiter receives a timeout arg, then it is expected to wait for
approximatedly this time (in seconds). To achieve this, the timeout used for the
sockets now are calculated from the timeout argument divided by the number of
tries.

This is not much a change to the up waiter that, in practice, was respecting a
timeout of timeout*tries. The down server, however, did not respected any
timeout. Since the value give to it was merely the socket timeout, and the
default behavior is to have the connection accepted while waiting, then it ran
another connection immediately after connecting - which was a quite short time
interval. To remedy this, the function sleeps for the socket timeout after a
successful connection.