#15 Make the test more reliable.
It costs a bit more of milliseconds (that are more important than one can
think...) but at least the test will break less. In an ideal world where all of
this is used, this will save *seconds* of attention of a developer running the
tests or watching a CI server. Like I've said, in an ideal world.