#29 Formatting.
- Now we get the caller at the finder constructor. Getting it from the other
method would result on calling sys._getframe(2). It is not easy to grasp what
this "2" means, and it is clearly a fragile call (what if we called
add_doctest() from inside yet another method?). The initializer is the only
place where the position of the calling method in the frame stack is known.

- We have a less compact add_doctest() method. It used to be a bit cryptic.

- Using herestrings in the tests.