#48 Change is_adoptable() results.
Originally, is_adoptable() checked whether its argument were builtin. We did it
because some builtins (e.g. dict) have a read-only __module__ attr.

Bug #48 showed the short-sighedtness of this logic. It doesn't really matter if
the object is built in - if the __module__ attr is rewritable, then it is
adoptable. So we changed is_adoptable() semantics: an object is adoptable if
it has a writable __module__ attr.

While the conceptual change is simple, the implementation change was a bit
complicated. We created a function, is_module_rewritable(), to check if
__module__ can be overridden in a specific object. While this function fails
when the object has no __module__ itself, it is important to have it separated
because it is used by the new AdoptionException messages.

Also, to check whether __module__ is writable on a specific object, it is
important to check if the object is a method. If it is so, we have to get the
method function instead. We used to do it in adopt(), but now we have to do it
in two places, so we extracted a new function: get_adoptable_value(). If its arg
is a method, it return the method's function; if no, it returns the object
itself.