#9 Make ugly.test.TestFinder callable from setup.py doctest.
NOW IT IS SENDING THE FIX. I REALLY NEED TO SLEEP. THAT SAID...

This is a weird behavior: if we call readme.rst through "python setup.py test",
it will fail. Apparently, the module where the doctest is executed has no name
and get_caller_module() fails. This is the error we get:

----------------------------------------------------------------------
File "/home/adam/software/ugly/ugly/../readme.rst", line 452, in readme.rst
Failed example:
    with ugly.module.installed_module('a', defs=[add]) as a,\
            ugly.module.installed_module('ta', defs=[TestAdd]) as ta:
        finder = ugly.finder.TestFinder(a, ta)
        import sys
        runner = unittest.TextTestRunner(stream=sys.stdout)
        runner.run(finder) # doctest: +ELLIPSIS
Exception raised:
    Traceback (most recent call last):
      File "/usr/lib/python2.7/doctest.py", line 1315, in __run
        compileflags, 1) in test.globs
      File "<doctest readme.rst[77]>", line 3, in <module>
        finder = ugly.finder.TestFinder(a, ta)
      File "/home/adam/software/ugly/ugly/finder/finder.py", line 124, in __init__
        caller_module = get_caller_module()
      File "/home/adam/software/ugly/ugly/module/module.py", line 442, in get_caller_module
        return importlib.import_module(frame.f_globals['__name__'])
    KeyError: '__name__'
----------------------------------------------------------------------
F

I do not know what is going on here. It is not the first time "setup.py test"
bite us but we really want to move on. Ideally, we would write a test to
reproduce the bug and then research it for the proper solution. However, since
we already spent too much time with ugly without releasing, we are going to add
this kludge.