#53 Count the frames to discover failure origin.
This is not really very clear. Suppose we import a module "foo" that, inside
it, does raise ImportError. In this case, the list of frames has three items:
inelegant.finder.finder, importlib and foo.

The imported module exists, but is failing with ImportError. In these cases, we
want rather to have TestFinder's search interrupted by re-raising the
exception. If we do not do that, either the search will fail with IOError (when
trying to read a non-existent "foo" file) or, worse, such a file will exist,
but may not contain doctests. In this case, we would not see tests or doctests
from "foo" being executed and there would be no clue of what is going on.

Now, suppose we import a module "bar" that does not even exist. In this case,
the list of frames will have only two items: inelegant.finder.finder and
importlib.

This case is different: since no such module exists, we may be trying to import
a file with doctests. In this case, we want to proceed with the search.

What we conclude from it? When a module exists but fails (and we want to
re-raise the error), the list of frames is greater than two, because the
exception came either from the faulty module itself - len(frames) == 3 - or
from a function called from it - len(frames) > 3. In this case, the exception
should be raised again. If the list has size 2, then the exception should be
ignored.