#57 Move tests from their packages to inelegant.test.
Mostly, we just moved files and renamed references. Some changes were made in
some places (e.g. the inelegant/test/__main__.py file) but they are a bit
minor.

Notably, we had to define the test loader in setup.py. Without that, waiter and
server doctests failed. The reasons are still unkown, but it is clear that
inelegant.net deserves some revision.