#81 Test: installed_module() restores previous module.
The purpose of installed_module() is to add a module to the system modules and
then remove it once installed_module's context is done. However, what should
happen if there still was a module with the same name of the one geing
installed. Well, it should be resotred. With this change, installed_module()
will restore the previous module if it exists.