#83 If enabled by toggle, use path as a prefix to be added to name.
By default, "path" argument should point directly to the resource to be
created. Yet, up to inelegant 0.1.0, this argument was used solely as a prefix
to be prepended to the "name" argument (very much like the "where" argument is
used today.) If you are dependent on this old behavior, it can be enabled via
the available_resource_uses_path_as_where toggle at inelegant.module.