#86 Adapt constructor to receive either function or file.
We want to use redirect_stdeout() as a decorator itself instead of using its
instance. This way, we can drop the parenthesis at the decorator. Instead of
writing

    @redirect_stdout(StringIO())
    def f(): print 'ok'

or

    @redirect_stdout()
    def f(): print 'ok'

we could then write

    @redirect_stdout
    def f(): print 'ok'

Yet, this means that the redirect_stdout() initializer can receive either a
function or a file-like object as its argument. If it receives a function, its
will decorate it; if it receives a file, it will be instantiated as a context
manager (or even as a decorator).

In this commit, we decide whether the first argument is a file or a callable.
That is not enough by itself to support the syntax we want to support but it is
the first step and can be a bit confusing by itself, so we are commiting it
alone.