#86 If no function is given to __call__ use the one given to __init__.
redirect_stdout() initializer can be either called with a file-like object as
its argument - e.g.

    o = StringIO()
    with redirect_stdout(o):
        print 'test'

or

    o = StringIO()
    @redirect_stdout(o)
    def f():
        print 'test'

- or be called with a callabe as its argument - as in

    @redirect_stdout
    def f():
        print 'test'

It means we have to make a decision when calling __call__: either it has to
decorate a function, or behave as a function itself. That is the decision we
make here: if a function was given to __init__, __call__ should be the
decorated function. Otherwise, call should just decorate the function it is
(hopefully) receiving.