#86 Refactor: use RedirectContextManager as decorator.
Since RedirecContextManager is a decorator by itself, we can use it instead of
redirect_decorator(). This renders redirect_decorator() unnecessary - again!

While separanting the concerns proved useful to some refactoring (and to fix a
bug with decorated bound methods) it had a problem: we would create a context
manager inside the __call__ of another context manager. This current
refactoring removes this double instantiation.

Overall, after all other changes, to merge both features makes sense. What
stays isolated is the decision about whether the instance will be a context
manager or a decorator. This one we make in the function.

Also, there is the gain that now the class has a beautiful, CamelCase name :)