#86 Refactor: give a more generic name to the redirect class.
This class can replace any attriute in any object. By renaming it from
RedirectContextManager to TemporaryAttributeReplacer, as well as its
constructor arguments and attributes, its generic applicability is made
explicit.

Given this new approach, there is no sense on defining the default value to
replace the old one in the replacer itself. I mean, when it replaced file-like
objects only, we could use StringIO; now it can replace anything. To make
things simpler, we moved the check for None to the functions that instantiate
the class.