#88 Remove support for using redirect_stdout()/redirect_stderr() as suppressors.
Not redirect_stdout() and redirect_stderr() are not decorators by themselves.
This way, they cannot be used for suppressing output anymore. For that, use the
suppressors suppress_stdout() and suppress_stderr().

The redirect_*() return values are still decorators, however, so they still can
be used to send the output to other files or file-like objects.